add travis-ci build script
In order to build for iOS without using a signing-key, we need to
target the iphone simulator, so let's add a config file for doing
just this.

In a future version of Uno, it'll be possible to do this from the
command-line, but for now this seems to be the only way to do this.